Tentative localization fix, updating position and direction now threaded

+++ map.h
@@ -1,6 +1,6 @@
 #include "map.c"
 
-void update_map (int x, int y, int dir, int values, int *obstacles, int *angles);
+void update_map (int x, int y, float dir, int values, int *obstacles, int *angles);
 
 void map_print(int startX, int startY, int endX, int endY);
